fix(usage): act on the export review
Seven of the eight items; what changed and why, in the reviewer's order.

1. Two toasts per failed export. The comment claiming the global interceptor
   cannot see a blob error was simply wrong: plugin-request calls
   errorConfig.errorHandler before it rejects, so every failure showed the
   localized, actionable message next to axios's own "Request failed with
   status code 422" — and a deliberately-silent estimate still popped the
   technical one. The four export endpoints now set skipErrorHandler and the
   comment says what actually happens.

   Taken together with the review's dispatch note: the token helpers now take
   the url like the resource ones, so there is one pair of functions, and
   `isTokenEndpoint = exportUrl === USAGE_BREAKDOWN_EXPORT` — dispatching on a
   string comparison — is gone along with the duplicates.

2. A preview with only a row-number column. The columns come from the
   estimate, so until it lands there is nothing to render; the table's spinner
   watched only the preview request, and the full-range COUNT is normally the
   slower of the two. `estimating` now feeds the spinner.

   Silencing the toast in (1) made the failure case worse, not better: the
   estimate could fail with no signal at all and leave an empty table that
   reads as "no data". So the hook reports `estimateFailed` and both dialogs
   say the size is unknown while leaving the download available — the server
   enforces the limits regardless.

3. The resource dialog re-sized the export on every page change: fetchEstimate
   rode along at the end of fetchPreview, whose effect depends on pageParams.
   The estimate covers the whole filtered range, so paging cannot alter it —
   each click was another aggregate query over the same rows. It now has its
   own effect keyed on the filters, which is where the Tokens dialog already
   had it.

4. The three "Export Table Data" entries had no feedback, which is exactly
   backwards: they are the ones that size and then stream with no dialog in
   between, under a hint that says "this may take a while". The dropdown
   trigger is the only visible control, so it takes the loading state.

5. Modal.confirm is the one static antd call this PR added, and static calls
   render outside the app's ConfigProvider — a light dialog with an
   untranslated "Cancel" in dark mode. Adds the antd <App> bridge (with
   component={false}, so no DOM node) and switches to modal.confirm from
   App.useApp. message.* stays as-is; it is already the repo's convention.

6. fileName / sheetName were declared, destructured and passed but unread
   since the server took over naming — removed, along with the dateSuffix and
   exportConfig fields that fed them. The resource tabs now import the export
   endpoints from apis/resource instead of re-typing the same paths in a local
   constant.

7. useExportPreviewColumns and useExportPreviewLayout move to hooks/ per
   CLAUDE.md, and split: columns change with the exported schema, the height
   math changes with the modal's chrome, and the value readers change with a
   breakdown response — three reasons to edit, now three files. Both sticky
   headers use <Flex vertical gap={12}> instead of inline flex properties.

8. The fallback filename hardcoded .csv while the request asks for xlsx, so an
   unreadable Content-Disposition handed back an xlsx (or a zip) under a name
   that opens as neither; it now derives from the format the estimate
   announced. `split` is declared on UsageExportRequest and organization_kind
   on the filter identity — both were being sent or read, and only compiled
   because the payloads were typed any. Dropped the two RESOURCE_FIELD_BY_KEY
   entries for groupings no tab exports.

   available / reason really are unconsumed. Kept, with the comment corrected
   to say why: the multi-sheet entries build their sheet list from the same
   visibility gate the tab strip uses, so an unavailable sheet is never
   requested — but per-sheet reporting is what lets a future dialog name the
   offending table instead of failing whole.

NOT done: the preview's PREVIEW_CHROME_HEIGHT / SUGGESTIONS_HEIGHT constants.
The reviewer is right that a measured alert height (or a flex content box)
would survive a translation that wraps; it needs the table body sized from a
resize observer rather than arithmetic, which is a bigger change than the rest
of this and is better done on its own.
